pyproject.toml
[project]
name = "biv-hack-infer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fire>=0.7.0",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "pymorphy3>=2.0.2",
    "pyspellchecker>=0.8.1",
    "scikit-learn>=1.5.2",
    "tqdm>=4.67.0",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
]
